Clear Theme in System UI
When SystemUI first starts an android theme object is created
which contains references to the currently applied CM theme.

When the theme is changed to system, the theme remains in memory
with old references to the prior applied theme. (See mPackages in ResTable::Theme)

This patch introduces a recreate theme method into Context so that
SystemUI can recreate its own theme object.

Change-Id: I086a76afa6f456a69c0390573bc8af2eafa4fb4e