Compile app theme in background
When a lot of themes are installed app installation takes a lot
of time blocking UI. In some cases phone gets rebooted because
of that (Watchdog kills PackageManager for blocking UI for too long).

This commit changes that - only current app theme is compiled on UI
thread, rest is compiled in background (like during theme installation).

Change-Id: Ied2ded26172b1b140e538b3906b25db808e05828
Signed-off-by: AudioGod <audiogod@sungsonic.co>